apps/slack-bot/pyproject.toml
[project]
name = "slack-bot"
version = "0.1.0"
description = "Slack bot for two-way sync with AI agents via Restack"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "slack-bolt>=1.21.2",
    "slack-sdk>=3.33.4",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.0",
    "restack-ai>=0.0.114",
    "asyncpg>=0.29.0",
    "watchfiles>=0.24.0",
    "starlette>=0.45.0",
    "uvicorn>=0.34.0",
]

[project.scripts]
dev = "src.server:dev"
start = "src.server:start"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "ruff>=0.8.5",
]
